+++ Assets/DOTS/Components/MetroLineComponent.cs
@@ -4,7 +4,7 @@ namespace DOTS.Components
 {
     public struct MetroLineComponent : IComponentData
     {
-        public int MetroLineID;
+        public byte MetroLineID;
         public Entity railPrefab;
     }
 }
